+++ src/field/mod.rs
@@ -39,8 +39,9 @@ pub trait Field {
     /// Advances the position `r` and normalized momentum `u`
     /// of a particle with charge to mass ratio `rqm`
     /// by a timestep `dt`, returning a tuple of the new
-    /// position and momentum
-    fn push(&self, r: FourVector, u: FourVector, rqm: f64, dt: f64) -> (FourVector, FourVector);
+    /// position and momentum, as well as the change in
+    /// lab time (which may differ from `dt`)
+    fn push(&self, r: FourVector, u: FourVector, rqm: f64, dt: f64) -> (FourVector, FourVector, f64);
 
     /// Checks to see whether an electron in the field, located at
     /// position `r` with momentum `u` emits a photon, and if so,
@@ -51,13 +52,14 @@ pub trait Field {
     /// Checks to see if an electron-positron pair is produced by
     /// a photon (position `r`, normalized momentum `ell`), returning the
     /// probability that it occurs in the specified interval `dt` and,
-    /// if so, the momentum of the electron and positron.
+    /// if so, the fraction of the photon that decays and the
+    /// the momentum of the electron and positron.
     ///
     /// A non-unity `rate_increase` makes pair creation more probable
     /// by the given factor, increasing the statistics for what would
     /// otherwise be a rare event. The probability returned is *not*
     /// affected by this increase.
-    fn pair_create<R: Rng>(&self, r: FourVector, ell: FourVector, dt: f64, rng: &mut R, rate_increase: f64) -> (f64, Option<(FourVector, FourVector)>);
+    fn pair_create<R: Rng>(&self, r: FourVector, ell: FourVector, dt: f64, rng: &mut R, rate_increase: f64) -> (f64, f64, Option<(FourVector, FourVector)>);
 }
 
 #[cfg(test)]
@@ -78,7 +80,7 @@ mod tests {
 
         // ponderomotive solver
         let dt = 0.5 * 0.8e-6 / (SPEED_OF_LIGHT);
-        let mut pond = (r, u);
+        let mut pond = (r, u, dt);
         for _i in 0..(20*2*2) {
             pond = laser.push(pond.0, pond.1, ELECTRON_CHARGE / ELECTRON_MASS, dt);
         }
@@ -87,7 +89,7 @@ mod tests {
         // Lorentz force solve
         // ponderomotive solver
         let dt = 0.01 * 0.8e-6 / (SPEED_OF_LIGHT);
-        let mut lorentz = (r, u);
+        let mut lorentz = (r, u, dt);
         for _i in 0..(20*2*100) {
             lorentz = fast_laser.push(lorentz.0, lorentz.1, ELECTRON_CHARGE / ELECTRON_MASS, dt);
         }
@@ -115,7 +117,7 @@ mod tests {
 
         // ponderomotive solver
         let dt = 0.5 * 0.8e-6 / (SPEED_OF_LIGHT);
-        let mut pond = (r, u);
+        let mut pond = (r, u, dt);
         for _i in 0..(20*2*2) {
             pond = laser.push(pond.0, pond.1, ELECTRON_CHARGE / ELECTRON_MASS, dt);
         }
@@ -124,7 +126,7 @@ mod tests {
         // Lorentz force solve
         // ponderomotive solver
         let dt = 0.01 * 0.8e-6 / (SPEED_OF_LIGHT);
-        let mut lorentz = (r, u);
+        let mut lorentz = (r, u, dt);
         for _i in 0..(20*2*100) {
             lorentz = fast_laser.push(lorentz.0, lorentz.1, ELECTRON_CHARGE / ELECTRON_MASS, dt);
         }